requirements-dev.txt
# Development dependencies - exact versions for reproducibility
black==25.1.0
flake8==7.3.0
isort==6.0.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-asyncio==1.1.0
pytest-homeassistant-custom-component==0.13.272
mypy==1.17.1
pre-commit==4.3.0
types-requests==2.32.4.20250809
types-cryptography==3.3.23.2
bandit==1.8.0
safety==3.2.8

# Additional testing utilities
responses==0.25.8
aioresponses==0.7.7

# MQTT testing
paho-mqtt>=1.6.0

# Mock libraries for testing Dyson devices
pytest-mock==3.14.0
